Move refreshWalletDataPublicKey stuff from fio to fioWalletsData
