+++ client/src/pages/FioTokensRequestPage/index.ts
@@ -3,7 +3,7 @@ import { createStructuredSelector } from 'reselect';
 
 import { compose } from '../../utils';
 
-import { refreshWalletDataPublicKey } from '../../redux/fio/actions';
+import { refreshWalletDataPublicKey } from '../../redux/fioWalletsData/actions';
 import { createContact, getContactsList } from '../../redux/contacts/actions';
 
 import {
