+++ client/src/redux/fio/actions.ts
@@ -190,10 +190,3 @@ export const updatePublicAddresses = (
   fioAddress,
   updPublicAddresses,
 });
-
-export const REFRESH_WALLET_DATA_PUBLIC_KEY = `${prefix}/REFRESH_WALLET_DATA_PUBLIC_KEY`;
-
-export const refreshWalletDataPublicKey = (publicKey: string) => ({
-  type: REFRESH_WALLET_DATA_PUBLIC_KEY,
-  publicKey,
-});
