+++ client/src/redux/fio/reducer.ts
@@ -376,13 +376,4 @@ export default combineReducers({
         return state;
     }
   },
-  walletDataPublicKey(state: string = '', action) {
-    switch (action.type) {
-      case actions.REFRESH_WALLET_DATA_PUBLIC_KEY: {
-        return action.publicKey;
-      }
-      default:
-        return state;
-    }
-  },
 });
