+++ client/src/redux/fio/selectors.ts
@@ -23,8 +23,6 @@ export const feesLoading = (state: ReduxState) => state[prefix].feesLoading;
 export const nftSignatures = (state: ReduxState) => state[prefix].nftList;
 export const mappedPublicAddresses = (state: ReduxState) =>
   state[prefix].mappedPublicAddresses;
-export const walletDataPublicKey = (state: ReduxState) =>
-  state[prefix].walletDataPublicKey;
 
 export const currentWallet = (
   state: ReduxState,
