+++ client/src/redux/fioWalletsData/selectors.ts
@@ -2,3 +2,5 @@ import { ReduxState } from '../init';
 import { prefix } from './actions';
 
 export const fioWalletsData = (state: ReduxState) => state[prefix].walletsData;
+export const walletDataPublicKey = (state: ReduxState) =>
+  state[prefix].walletDataPublicKey;
